backend/requirements.txt
beautifulsoup4
dataclasses-json
faiss-cpu
Flask
Flask-Cors
Flask-RESTful
Flask-SQLAlchemy
frozenlist
fuzzywuzzy
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
lxml
openai
openapi-schema-pydantic
orjson
outcome
overrides
packaging
pydantic
pydantic_core
python-dateutil
python-dotenv
python-json-logger
SQLAlchemy
stack-data
tenacity
terminado
tiktoken
youtube-transcript-api
requests
pytz
PyPDF2
pdf2image